+++ src/app/view.rs
@@ -93,8 +93,8 @@ impl OpenCADStudio {
                         let aspect = if vh > 0.0 { vw / vh } else { 1.0 };
                         let half_h = cam.ortho_size();
                         let half_w = half_h * aspect;
-                        let tx = cam.target.x;
-                        let ty = cam.target.y;
+                        let tx = cam.target.x as f32;
+                        let ty = cam.target.y as f32;
                         drop(cam);
                         grips_to_screen_paper(&tab.selected_grips, tx, ty, half_w, half_h, bounds)
                     } else {
